Renames folder to reflect usage
