+++ apps/judicial-system/web/src/routes/Shared/CaseOverview/SignedVerdictOverview.tsx
