+++ apps/judicial-system/web/src/routes/Shared/CaseOverview/courtRecordSignatureConfirmationGql.ts
